Move GraphQL set up to App.js
