+++ src/index.js
@@ -1,9 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import { ApolloClient } from 'apollo-client';
-import { ApolloProvider } from 'react-apollo';
-import { InMemoryCache } from 'apollo-cache-inmemory';
-import { HttpLink } from 'apollo-link-http';
 
 // eslint-disable-next-line
 import tnt from 'tntvis';
@@ -13,8 +9,9 @@ import utils from 'tnt.utils';
 import rest from 'tnt.rest';
 
 import tooltip from './utils/tooltip';
-import App from './App';
 import * as serviceWorker from './serviceWorker';
+
+import App from './public/App';
 
 // styling
 import 'dc/dc.min.css';
@@ -25,19 +22,7 @@ import './index.scss';
 // window.tnt = tnt;
 window.tnt.tooltip = tooltip;
 
-const client = new ApolloClient({
-  link: new HttpLink({
-    uri: 'https://platform-api.now.sh/graphql',
-  }),
-  cache: new InMemoryCache(),
-});
-
-ReactDOM.render(
-  <ApolloProvider client={client}>
-    <App />
-  </ApolloProvider>,
-  document.getElementById('root')
-);
+ReactDOM.render(<App />, document.getElementById('root'));
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
